+++ app/controllers/api/CHANGELOG.md
@@ -1,5 +1,12 @@
+### 2.2.0 (2013-05-20)
+* Added Audio API endpoint
+* Added Audio array to Content object
+* Updated documentation in the README
+
+
 ### 2.1.1 (2013-05-07)
 * Cache content objects.
+
 
 ### 2.1.0 (2013-04-22)
 * Added assets array (4 sizes: thumbnail, small, large, full), each with URL and dimensions. Assets array will always be present, even if there are no assets (blank array.)